SaNN running on PROTEINS

+++ topobenchmarkx/nn/wrappers/simplicial/sann_wrapper.py
@@ -60,7 +60,7 @@ class SANNWrapper(AbstractWrapper):
         x_out = self.backbone(x_all)
 
         model_out = {
-            "labels": batch.y.type(torch.float32),
+            "labels": batch.y.type(torch.long),
             "batch_0": batch.batch_0,
             "batch_1": batch.batch_1,
             "batch_2": batch.batch_2,
